Move update_* functions to update.c.
